indentation, move and removal of some files
