--- src/Rental.js
@@ -1,15 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.rent = void 0;
-exports.rent = [
-    {
-        idStudent: 1,
-        idLivros: 1,
-        date: new Date(2024, 7, 8)
-    },
-    {
-        idStudent: 3,
-        idLivros: 4,
-        date: new Date(2024, 9, 12)
-    }
-];
