--- src/Student.js
@@ -1,21 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.students = void 0;
-exports.students = [
-    {
-        idStudent: 1,
-        nameStudent: "Junior",
-        borrowedBooks: 2
-    },
-    {
-        idStudent: 2,
-        nameStudent: "Lucas",
-        borrowedBooks: 3
-    },
-    {
-        idStudent: 3,
-        nameStudent: "João",
-        borrowedBooks: 4
-    }
-];
-//console.log(students)
