--- src/interAluguel.js
@@ -1,15 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.rent = void 0;
-exports.rent = [
-    {
-        iD_Student: 1,
-        iD_Livros: 1,
-        date: new Date(2024, 7, 8)
-    },
-    {
-        iD_Student: 3,
-        iD_Livros: 4,
-        date: new Date(2024, 9, 12)
-    }
-];
