--- src/teste.js
@@ -1,21 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const Rental_1 = require("./Rental");
-const Book_1 = require("./Book");
-const Student_1 = require("./Student");
-const functions_1 = require("./functions");
-console.log("---------------------------------------------- Livros disponiveis:------------------------------------------------");
-console.log((0, functions_1.getAvailableBooks)());
-console.log("---------------------------------------------- Titulos de todos os livros: ---------------------------");
-console.log((0, functions_1.getBookTitles)());
-console.log("---------------------------------------------- Numeros totais de livros por autor:-----------------------------------");
-console.log(`Total de Livros: `, functions_1.chooseAuthor);
-console.log("---------------------------------------------- Livros de forma Ordenados:---------------------------------------------");
-console.log(functions_1.orderedBooks);
-console.log("------------------------------------------------  ALUGUEL -------------------------------------");
-console.log(functions_1.resultOfRent);
-console.log(`Quantidades desse livro: ${Book_1.books.find(books => books.idBook === 2)?.quantity}`);
-console.log(`Quantidades de livros ja alugados deste aluno: ${Student_1.students.find(students => students.idStudent === 1)?.borrowedBooks}`);
-console.log(`Horário: ${functions_1.dateOfRent}`);
-console.log("------------------------------------------------ REGISTRO DO ALUGUEL ----------------------------");
-console.log(Rental_1.rent);
